Fix #2 by requiring a certain history.length + go forward after back
The required length depends on how the tab was opened (directly (through about:blank) or through about:newtab).

Also makes it automatically press the forward button after (or actually right before) closing the tab, which makes restoration of tabs not lose the BACK_CLOSE state even if it's the last one.